actions/requirements.txt
boto3==1.20.8
Pillow==8.4.0
rasa_sdk==2.8.2
requests==2.25.0
schedule==1.1.0
